use router to navigate between Dashboard and EspIdf installation
